# docs/trading_alerts_openapi_compress.yaml
openapi: 3.0.3
info:
  title: Trading Alerts SaaS API
  version: 7.1.0
  description: |
    Commercial SaaS platform for trading alerts and signal generation with real-time chart visualization.
    **Stack:** Next.js 15 + Flask MT5 + Dual Payment (Stripe/dLocal)
    **Auth:** NextAuth.js v4 (Google OAuth + Email/Password, JWT sessions)
    **Tiers:** FREE (5 symbols × 3 timeframes = 15 combos), PRO (15 symbols × 9 timeframes = 135 combos)
    **Payment:** Stripe (international, auto-renewal, 7-day trial), dLocal (8 markets: IN, NG, PK, VN, ID, TH, ZA, TR, manual renewal)
  contact:
    name: API Support
    email: support@tradingalerts.com
  license:
    name: Proprietary
    url: https://tradingalerts.com/license
servers:
  - url: https://api.tradingalerts.com
    description: Production server
  - url: https://staging-api.tradingalerts.com
    description: Staging server
  - url: http://localhost:3000
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/auth/forgot-password:
    post:
      tags:
        - Authentication
      summary: Request password reset
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
              required:
                - email
      responses:
        '200':
          description: Password reset email sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /api/auth/reset-password:
    post:
      tags:
        - Authentication
      summary: Reset password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                password:
                  type: string
                  minLength: 8
              required:
                - token
                - password
      responses:
        '200':
          description: Password reset successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ============================================
  # OAUTH AUTHENTICATION ENDPOINTS
  # ============================================
...
  /api/checkout:
    post:
      tags:
        - E-commerce
      summary: Create checkout session for PRO upgrade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tier:
                  type: string
                  enum: [PRO]
                billingCycle:
                  type: string
                  enum: [monthly]
                  default: monthly
                  description: Only monthly subscriptions available (no annual plans yet)
              required:
                - tier
      responses:
        '200':
          description: Checkout session created
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                    format: uri
                    description: Stripe checkout URL
        '400':
          description: Invalid tier or already subscribed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/notifications:
    get:
      tags:
        - Notifications
      summary: Get user notifications
      parameters:
        - name: unread
          in: query
          schema:
            type: boolean
          description: Filter unread only
      responses:
        '200':
          description: List of notifications
          content:
            application/json:
              schema:
                type: object
                properties:
                  notifications:
                    type: array
                    items:
                      $ref: '#/components/schemas/Notification'
  /api/notifications/{id}/read:
    put:
      tags:
        - Notifications
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Notification marked as read
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from login (includes tier information)
...
  schemas:
...
    NotificationType:
      type: string
      enum:
        - ALERT
        - SYSTEM
        - BILLING
        - SECURITY
        - FEATURE
...
    Notification:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        type:
          $ref: '#/components/schemas/NotificationType'
        title:
          type: string
        message:
          type: string
        data:
          type: object
          nullable: true
        isRead:
          type: boolean
        readAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time

    # ============================================
    # COMMON SCHEMAS
    # ============================================
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
          nullable: true
        details:
          type: object
          nullable: true
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
          nullable: true
...